+++ haskell/src/Lorentz/Contracts/StablecoinFA1_2.hs
@@ -8,28 +8,29 @@ module Lorentz.Contracts.StablecoinFA1_2
   , Parameter(..)
   , stablecoinFA1_2Contract
 
+  , GetCounterParam
+  , GetDefaultExpiryParam
+
   -- * TZIP-16
-  , metadataMap
   , metadataJSON
   ) where
 
-import qualified Data.Aeson as J
-import qualified Data.ByteString.Lazy as BSL
 import Data.FileEmbed (embedStringFile)
-import qualified Data.Map as Map
 import Fmt (pretty)
 
+import Data.Version (showVersion)
 import Lorentz as L
 import qualified Lorentz.Contracts.Spec.ApprovableLedgerInterface as AL
+import Lorentz.Contracts.Spec.TZIP16Interface
+  (Metadata(..), MetadataMap, ViewImplementation(..))
 import Michelson.Runtime (parseExpandContract)
 import qualified Michelson.Untyped as U
+import Morley.Metadata (ViewCode(..), mkMichelsonStorageView)
 
-import Lorentz.Contracts.Spec.TZIP16Interface
-  (Metadata(..), MetadataMap, SomeMichelsonStorageView(..), ViewImplementation(..),
-  mkMichelsonStorageView)
-import qualified Lorentz.Contracts.Spec.TZIP16Interface as TZ (View(..))
+import qualified Lorentz.Contracts.Spec.TZIP16Interface as TZ
 import qualified Lorentz.Contracts.Stablecoin as S
 import Lorentz.Contracts.StablecoinPath (stablecoinFA1_2Path)
+import Paths_stablecoin (version)
 
 data Storage = Storage
   { sDefaultExpiry :: S.Expiry
@@ -70,8 +71,8 @@ data Parameter
   | Change_master_minter S.ChangeMasterMinterParam
   | Change_pauser S.ChangePauserParam
   | Configure_minter S.ConfigureMinterParam
-  | Get_counter S.GetCounterParam
-  | Get_default_expiry S.GetDefaultExpiryParam
+  | Get_counter GetCounterParam
+  | Get_default_expiry GetDefaultExpiryParam
   | Mint S.MintParams
   | Pause
   | Permit S.PermitParam
@@ -86,6 +87,10 @@ data Parameter
 
 instance ParameterHasEntrypoints Parameter where
   type ParameterEntrypointsDerivation Parameter = EpdRecursive
+
+type GetCounterParam = View () Natural
+
+type GetDefaultExpiryParam = View () S.Expiry
 
 -- | Parse the metadata registry contract.
 stablecoinFA1_2Contract :: U.Contract
@@ -103,50 +108,30 @@ stablecoinFA1_2Contract =
       |]
     )
 
-metadataMap :: MetadataMap BigMap
-metadataMap = BigMap $ Map.fromList
-  [ (mempty, encodeUtf8 @Text "tezos-storage:metadataJSON")
-  , ([mt|metadataJSON|], BSL.toStrict (J.encode metadataJSON))
-  ]
-
 metadataJSON :: Metadata (ToT Storage)
 metadataJSON =
-  S.metadataJSON
-    { mName = Just "stablecoin FA1.2"
-    , mInterfaces = [ "TZIP-7", "TZIP-17" ]
-    , mViews =
-        [ getDefaultExpiryView
-        , getCounterView
-        ]
-
-    -- NOTE: Because we're storing the metadata in the contract's storage,
-    -- the storage may now exceed the storage size limit.
-    -- See: https://buildkite.com/serokell/stablecoin/builds/1150#eb990412-157b-4a96-92d8-3fd39d954369/65-261
-    --
-    -- As a temporary measure, we're removing the "homepage" and "source"
-    -- info to make the storage a little bit smaller, just enough to make it
-    -- possible to originate the contract.
-    --
-    -- TODO: once we move the metadata to its own contract, we can
-    -- add this info back to the metadata.
-    , mHomepage = Nothing
-    , mSource = Nothing
-    -- , mSource = Just Source
-    --     { sLocation = "https://github.com/tqtezos/stablecoin/tree/v" <> toText (showVersion version) <> "/ligo/stablecoin/fa1.2"
-    --     , sTools = [ "ligo " ]
-    --     }
-    }
+  TZ.name "stablecoin FA1.2" <>
+  TZ.interfaces [ TZ.Interface "TZIP-007", TZ.Interface "TZIP-017" ] <>
+  TZ.views
+      [ getDefaultExpiryView
+      , getCounterView
+      ] <>
+  TZ.source
+    (TZ.Source
+      (Just $ "https://github.com/tqtezos/stablecoin/tree/v" <> toText (showVersion version) <> "/ligo/stablecoin/fa1.2")
+      [ "ligo " ]) <>
+  TZ.homepage "https://github.com/tqtezos/stablecoin/"
 
 getDefaultExpiryView :: TZ.View (ToT Storage)
 getDefaultExpiryView =
   TZ.View
     { vName = "GetDefaultExpiry"
     , vDescription = Just "Access the contract's default expiry in seconds"
-    , vPure = True
+    , vPure = Just True
     , vImplementations = one $
-        VIMichelsonStorageView $ SomeMichelsonStorageView $
-          mkMichelsonStorageView @Storage @() [] $
-            L.car # L.toField #sDefaultExpiry
+        VIMichelsonStorageView $
+          mkMichelsonStorageView @Storage  @Natural Nothing [] $ WithoutParam $
+            L.toField #sDefaultExpiry
     }
 
 getCounterView :: TZ.View (ToT Storage)
@@ -154,9 +139,9 @@ getCounterView =
   TZ.View
     { vName = "GetCounter"
     , vDescription = Just "Access the current permit counter"
-    , vPure = True
+    , vPure = Just True
     , vImplementations = one $
-        VIMichelsonStorageView $ SomeMichelsonStorageView $
-          mkMichelsonStorageView @Storage @() [] $
-            L.car # L.toField #sPermitCounter
+        VIMichelsonStorageView $
+          mkMichelsonStorageView @Storage @Natural Nothing [] $ WithoutParam $
+            L.toField #sPermitCounter
     }